Adding version 2.6.30.5 headers to base/lib/linux-arch/x86/include/asm   and deleting version 2.6.22.6 headers from base/lib/linux-headers/asm-i386:   types.h
Other changes to compile for CapROS.

+++ src/base/lib/linux-headers/asm-generic/int-ll64.h
@@ -20,8 +20,8 @@ typedef unsigned char __u8;
 typedef __signed__ short __s16;
 typedef unsigned short __u16;
 
-typedef __signed__ int __s32;
-typedef unsigned int __u32;
+typedef __signed__ /* int modified to match stdint.h */ long __s32;
+typedef unsigned /* int modified to match stdint.h */ long __u32;
 
 #ifdef __GNUC__
 __extension__ typedef __signed__ long long __s64;
